finish extras

+++ src/main/java/com/mindhub/homebanking/dtos/LoanApplicationDTO.java
@@ -1,10 +1,10 @@
 package com.mindhub.homebanking.dtos;
 public class LoanApplicationDTO {
     private Long id;
-    private int amount;
+    private double amount;
     private int payments;
     private String numberAccountDestini;
-    public LoanApplicationDTO(long id, int amount, int payments, String numberAccountDestini) {
+    public LoanApplicationDTO(long id, double amount, int payments, String numberAccountDestini) {
         this.id = id;
         this.amount = amount;
         this.payments = payments;
@@ -13,9 +13,7 @@ public class LoanApplicationDTO {
     public Long getId() {
         return id;
     }
-    public Integer getAmount() {
-        return amount;
-    }
+    public double getAmount() { return amount; }
     public Integer getPayments() {
         return payments;
     }
